Abstracted some logic
